+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v2.0.0-beta.1
 
+- Added
+  - [Auth] JWT now supports RS256 algorithm in addition to HS256 - [#84](https://github.com/Accenture/reactive-interaction-gateway/issues/84)
+
 ## v2.0.0-dev
 
 - Changed